Renamed currencySelector to coinSelector (had duplicate file/component names)
