+++ public/coinSelectorSVGs.tsx
